requirements.txt
gitpython
requests
beautifulsoup4
marker-pdf
python-dotenv